Add support for writeHead() in the compress middleware
compress looks at the headers (which haven't been written yet) in
order to make a decision as to whether or not the response will
be compressed.

This is an attempt at merging the headers that have been queued for
write (via setHeader()) and the ones which will be written via
writeHead() after compress has run.